pyproject.toml
[tool.poetry]
name = "sentei-pictures"
version = "0.1.0"
description = "Photo selection and compression utility"
authors = ["Claude Code Assistant <noreply@anthropic.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/user/sentei-pictures"
repository = "https://github.com/user/sentei-pictures"
keywords = ["photo", "image", "compression", "selection", "jpeg"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]
packages = [{include = "sentei_pictures", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
pytest = "^8.0.0"
pytest-cov = "^5.0.0"

[tool.poetry.scripts]
sentei = "sentei_pictures.cli.main:main"
sentei-reduce = "sentei_pictures.cli.reduce:main"
sentei-choice = "sentei_pictures.cli.choice:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
